+++ CHANGELOG.md
@@ -23,7 +23,8 @@ _ `_optional` subpackage for managing optional dependencies
 - Abstract `ContinuousNonlinearConstraint` class
 - `ContinuousCardinalityConstraint` class and corresponding uniform sampling mechanism
 - `register_hooks` utility enabling user-defined augmentation of arbitrary callables
-- Polars expression for `DiscreteSumConstraint`, `DiscreteProductConstraint` and `DiscreteExcludeConstraint`
+- Polars expressions for `DiscreteSumConstraint`, `DiscreteProductConstraint` and 
+  `DiscreteExcludeConstraint`
 
 ### Changed
 - Passing an `Objective` to `Campaign` is now optional